Missing namelist var and added add'l aero prof
Fixed missing namelist variable (CCN_FT), the CCN concentration in the
free troposphere for an ICCN_PROF option. Also added a new initial
aerosol profile that is constant in the boundary layer and uses a
exponential decay with a 3.5 km scale height above the boundary layer.
Added this since it seemed to possible be similarly realistic to the
other new option, but with fewer user settings.